pyproject.toml
[project]
name = "idaes-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "coolprop>=7.2.0",
    "idaes-pse>=2.8.0",
    "ahuora-compounds",
    "idaes-ui>=0.25.10",
    "ahuora-builder-types==0.1.4",
    "ahuora-builder==0.1.4",
    "mcp>=1.13.0",
]


[tool.uv.sources]
ahuora-compounds = { git = "https://github.com/waikato-ahuora-smart-energy-systems/PropertyPackages.git", rev = "v0.0.29" }
